Move websocket package under streaming
