+++ components/apimgt/org.wso2.carbon.apimgt.gateway/src/main/java/org/wso2/carbon/apimgt/gateway/handlers/streaming/websocket/WebSocketApiException.java
@@ -16,16 +16,12 @@
  * under the License.
  */
 
-package org.wso2.carbon.apimgt.gateway.handlers.websocket;
+package org.wso2.carbon.apimgt.gateway.handlers.streaming.websocket;
 
-public class WebSocketApiConstants {
+public class WebSocketApiException extends Exception {
 
-    WebSocketApiConstants() {
+    public WebSocketApiException(String message) {
+        super(message);
     }
 
-    public static final String WEBSOCKET_DUMMY_HTTP_METHOD_NAME = "WS";
-    public static final String WS_ENDPOINT_NAME = "WebSocketInboundEndpoint";
-    public static final String WS_SECURED_ENDPOINT_NAME = "SecureWebSocketEP";
-    public static final String URL_SEPARATOR = "/";
-    public static final String DEFAULT_RESOURCE_NAME = "/_default_resource_of_api_";
 }
